Add thread names and timing information to protobuf reports (#90)
* Add thread names and timing information to protobuf reports

Record thread names (or IDs); record "cpu: nanoseconds" in addition to
"samples: count" value types; and the timestamp and duration of the collected
profile; to protobuf reports.

This matches exactly the format of profiles produced by Golang's
`runtime/pprof` package. And, when inspected with `go tool pprof` default to
displaying seconds spent in a specific location, as opposed to samples (whose
counts depend on sampling frequency, requiring a lot of mental math to map to
actual time). It is of course still possible to view samples instead of
seconds, if wanted.

Signed-off-by: Alin Sinpalean <alin.sinpalean@gmail.com>

* Remove unnecessary default() call.

Signed-off-by: Alin Sinpalean <alin.sinpalean@gmail.com>

* Add thread names and timing information to protobuf reports

Record thread names (or IDs); record "cpu: nanoseconds" in addition to
"samples: count" value types; and the timestamp and duration of the collected
profile; to protobuf reports.

This matches exactly the format of profiles produced by Golang's
`runtime/pprof` package. And, when inspected with `go tool pprof` default to
displaying seconds spent in a specific location, as opposed to samples (whose
counts depend on sampling frequency, requiring a lot of mental math to map to
actual time). It is of course still possible to view samples instead of
seconds, if wanted.

Signed-off-by: Alin Sinpalean <alin.sinpalean@gmail.com>

* Remove unnecessary default() call.

Signed-off-by: Alin Sinpalean <alin.sinpalean@gmail.com>

* Move import of SystemTime under mod protobuf.

Signed-off-by: Alin Sinpalean <alin.sinpalean@gmail.com>